Enabling runImageTransitionIfCached now sets placeholder image (#293)
This change contains two important updates. The first is to set the placeholder image when loading from the cache with runImageTransitionIfCached enabled. This makes the behavior between loading from the cache and downloading from the network exactly the same. If you don’t do this, you can really tell the difference between cache and remote loads which we don’t want to do.

I also added a check to make sure the receipt hasn’t changed out from under us before running the transition. This is in regards to Issue #293 which I cannot validate. You must perform the queue hop to allow the image view to be redisplayed before running the animation. If you don’t do this, the image isn’t reset before running the transition which can lead to the initial display of the old image rather than the placeholder or no image. The queue hop MUST be done in order to ensure the image view is redrawn before the transition begins.